Modified G4. Added partial logic to ASTListener. Almost ready for Level0 parsing.
Modified more parts. Noticed the G4 was not matching the Fixtures structure. To make it eassier, use the predefined literals, operations and selectos.

Have to find a way to get Selector to work. Seems to be an abstract class rather then an implemented ASTNode. Might need to use PropertyName, but i find it quite weird to use PropertyName as a Selector element.